requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-import-export==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
PyYAML==5.3.1
sqlparse==0.4.1
tablib==2.0.0
uWSGI==2.0.19.1
xlrd==1.2.0
xlwt==1.3.0
